Move constants to Util.py
Had circular import trying to import util.py from messages. Move some optimizer code from util.py to
optimizers.py to break the circular import.

Then, move constants from constants.py to util.py.
