--- ssm/constants.py
@@ -1,3 +1,0 @@
-SEED = hash("ssm")
-LOG_EPS = 1e-16
-DIV_EPS = 1e-16
